requirements.txt
Appium-Python-Client==5.1.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
dotenv==0.9.9
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jira==3.8.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
